+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## 0.2.0 (unreleased)
 
+- Added option `-q`/`--quiet` to suppress output.
+
 
 ## 0.1.0 (2022-05-08)